Session Vault code moved to common.
