+++ wallet-sdk-core/src/androidMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/AccountMppSigner.kt
@@ -1,0 +1,16 @@
+package com.michaeltchuang.walletsdk.core.railmpp
+
+import com.algorand.algosdk.account.Account
+import com.algorand.algosdk.transaction.Transaction
+import com.algorand.algosdk.util.Encoder
+
+/** [AndroidMppWalletSigner] backed by an in-process algosdk [Account]. */
+class AccountMppSigner(
+    private val account: Account,
+) : AndroidMppWalletSigner {
+    override val address: String get() = account.address.toString()
+    override val authorizedSignerPublicKey: ByteArray get() = account.address.getBytes()
+
+    override suspend fun signTransaction(txn: Transaction): ByteArray =
+        Encoder.encodeToMsgPack(account.signTransaction(txn))
+}
